Correct formatting error in `equery` output
When using -F (and not --format), if an equals sign is included, it gets
added to the format string as a leading equals sign, which we had been
patiently chopping out. This changes it so that we don't include it in
the first place, saving a substitution operation.

Noticed during discussion of GitHub issue #49.