Rewrite report_time() with UTC everywhere
Remove the fighting with timezones, and instead just report the time in
UTC. If the report time is only for use in timewise graphs, and not for
geolocation of reports (e.g. categorising submissions by timezone), it's
not really needed.

I'd suggest that if the timezone *is* needed, it could be retrieved
using `POSIX::strftime` with the `%Z` specifier, and included in a
separate datum of the report.

This hopefully addresses issue #44 completely, if the above thinking is
accurate.