Don't retrieve lspci output on module load
Move the `get_lspci()` invocation from the code that runs when the
module file is imported to within the `get_hardware_info()` sub, which
is (presently) the only place it's actually needed.

This avoids running the code when the module is loaded, but methods that
require `lspci` output aren't called, such as simply showing the usage
help for the `funtoo-report` script.