Factor out arrayref transform into separate sub
Pull all of the hash transformation logic out of `get_filesystem_info`
and put it into a package subroutine instead, `transform_es_arrayref`,
which accepts the data structure by reference as a first parameter and
at hash key as a second: `name` in the current use case.

This function might actually be generically useful for other
ElasticSearch users. Maybe we should make another module out of it...

This commit does not (er, shouldn't) change the interface of the tool
nor its output in any way.