Replace `transform_es_arrayref` with simple method
I thought there would be more need for this subroutine then there
actually has been; only `get_filesystem_info` really needs it, because
we just build the rest of our data structure. Replace it with a much
simpler and ad-hoc processing of the `lsblk --json` output specifically.

We can always dig it back up later if it's needed, but I doubt we will.