Use magic #!perl shebang for binscript
When `funtoo-reporter` is built for installation by Makefile.PL, this
will use the full path to the version of Perl with which it was built
(or the specified Perl path), rather than relying on the /usr/bin/env
trick.

This is the part of the Perl distribution process that I am the least
confident will adapt correctly to Funtoo, but we'll see how it works
out.