Specify minimum Perl of v5.14
This is only a suggested starting point. I chose it because it's the
first stable release after HTTP::Tiny was made core, and the project
switched to using that for its data submission in commit 406e1666.

Perl v5.14 was first released in 2011, so it's pretty old. If necessary
we can very likely support Perl releases older than this, but this
seemed a sensible starting point.