Re-work get_all_installed_pkg()
Reworked get_all_installed_pkg() to get the data without the use of
equery and to make use of a pure hash datastructure. Now includes the
SLOT used for each package, though this addition has some speed
consequences due to the complexity of the hash.

Removed get_version_info() since it is likely redundant now that
searching can be performed easier on the ElasticSearch website.