Define a custom User-Agent for the HTTP submit
Just to be polite, set the User-Agent to something custom so that
requests from this script can be quickly identified in server logs.

In an Apache HTTPD log, the requests look something like this:

    192.0.2.1 - - [12/Mar/2018:09:58:07 +1300] "POST / HTTP/1.1" 201 23514 "-" "Funtoo::Report/1.4"

We can add more information to the user agent if we want, such as an
URL, but I think the name and version of the application is sufficient.