Reimplement get_kernel_info to read fewer files
Since we already know the names of the files in which we're interested,
and treat them all the same way if we can open them, there's no reason I
can see here to iterate through the entries in the /proc/sys/kernel
directory looking for them. We can just attempt to open the files by
name directly, saving some cycles and some subroutine complexity.