Add the beginnings of some POD documentation
Add POD to the end of both the funtoo-report script and the
Funtoo::Report module, with some recommended sections. At the moment
this is still very skeletal; in particular, documentation needs to be
added for each module subroutine under each heading, and some more
in-depth information about diagnostics.

This is related to GitHub issue #68.