Allow specifying alt path to config file
For the simplest way to get this working, add the expected `--config` or
`-c` option, and arrange for it to modify the value of Funtoo::Report's
$config_update variable if specified, by changing it to a package
variable (`our` vs `my`). Update documentation to match.

This is not an ideal implementation, but it's the simplest route to
getting what we need. Ideally what we should do is create instances of
the Funtoo::Report module with some very basic OOP, so that two modules
can exist in parallel with different configuration files. Such a change
requires consultation with the other developers.

For the moment, however, I just wanted something that would make
automated tests more straightforward to implement, and this should
accomplish that.

This implements the feature wanted for GitHub issue #82.