Add test suite infrastructure and very first test
Add a "t" directory to contain the automated tests for the distribution,
and put in a "require.t" test as its first entry, which only checks that
the module can be required (and hence used) without a fatal error.

This can be tested with:

    $ perl Makefile.PL
    $ make
    $ make test

The last line of your output should be:

    Result: PASS

Also specify in Makefile.PL that for the module build process, we now
require Test::More, which has been core since Perl v5.6.2.

This is the first commit in planned work for GitHub issue #97.