Move perltidy.t test to "xt" subdirectory
This is the name of the subdirectory that Module::Starter uses for
"developer tests", i.e. stuff that the developers may want to check but
that users won't necessarily care about, like linting, tidying,
documentation, and coverage tests. This means we don't require
Perl::Critic or Perl::Tidy just to test the usage of the module.

You can run these tests with:

    $ prove -lb xt

The "prove" script is from the Perl core, as far as I can tell.

I think there's a way to arrange a Makefile target for the developer
tests, but I don't know what it is just yet.