Dirty start to Changelog
This is just generated with:

    $ git log --pretty='format: - %s' --no-merges vA..vB --reverse

Where vA and vB are two adjacent version number tags, with headings and
text wrapping in the style of a Perl "Changes" file.

Ideally it should be cleaned up to remove other stuff that a user of the
module wouldn't care about, such as minor code refactoring that doesn't
change any user-visible functionality. I'll make a start on that in the
next couple of commits.

This is for GitHub issue #122.